ProtoViT/requirements.txt
# This file may be used to create an environment usin
# $ conda create --name <env> --file <this fil
# platform: linux
# _libgcc_mutex
# _openmp_mutex
absl-py
# blas
blessed
brotli-python
bzip2
ca-certificates
cairo
certifi
cffi
charset-normalizer
cleverhans
cloudpickle
contourpy
cryptography
cuda-cudart
cuda-cupti
cuda-libraries
cuda-nvrtc
cuda-nvtx
cuda-runtime
cycler
dbus
decorator
dm-tree
easydict
expat
ffmpeg
filelock
fontconfig
fonttools
freetype
fsspec
gast
gettext
giflib
glib
glib-tools
gmp
gmpy2
gnutls
gpustat
graphite2
gst-plugins-base
gstreamer
harfbuzz
hdf5
huggingface-hub
icu
idna
importlib-resources
intel-openmp
jasper
jinja2
joblib
jpeg
kiwisolver
krb5
lame
lcms2
ld_impl_linux-64
lerc
libblas
libcblas
libclang
libcublas
libcufft
libcufile
libcurand
libcusolver
libcusparse
libdeflate
libedit
libevent
libffi
libgcc-ng
libgfortran-ng
libgfortran5
libglib
libgomp
libiconv
libidn2
liblapack
liblapacke
libllvm11
libnpp
libnvjpeg
libopencv
libpng
libpq
libstdcxx-ng
libtasn1
libtiff
libunistring
libuuid
libwebp
libwebp-base
libxcb
libxkbcommon
libxml2
lz4-c
markupsafe
matplotlib
mkl
mkl-service
mkl_fft
mkl_random
mnist
mpc
mpfr
mpmath
mysql-common
mysql-libs
ncurses
nettle
networkx
ninja
nose
nspr
nss
numpy
numpy-base
nvidia-ml-py
opencv
openh264
openjpeg
openssl
packaging
pandas
pcre
pillow
pip
pixman
psutil
py-opencv
pycodestyle
pycparser
pyopenssl
pyparsing
pysocks
python
python-dateutil
python_abi
pytorch
pytorch-cuda
pytorch-mutex
pytz
pyyaml
qt
readline
regex
requests
safetensors
scikit-learn
scipy
setuptools
six
sqlite
sympy
tbb
tensorflow-probability
threadpoolctl
timm
tk
tokenizers
torchaudio
torchtriton
torchvision
tqdm
transformers
typing-extensions
tzdata
urllib3
wcwidth
wheel
xz
zipp
zlib
zstd
